Patch to vaf_ncm.py to add multithreading support wrapping around the pearson_def function for all pairs
Changed for loop for filtering vecA and vecB to shared features to a list comprehension

output_matrix loops through keylist instead of samples, and extracts keys from keylist instead of temp. Avoid redundant writes to output_matrix with quadratic scaling, which is the bottleneck when running on very large datasets. Implementation may not be ideal, but does correctly work for this use case.